requirements.txt
pydantic-settings
pydantic

pytelegrambotapi
google-genai